% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{exam}
\alias{exam}
\alias{loppukoe}
\title{The final exam for the R-course}
\usage{
exam(dob, ...)

loppukoe(dob, ...)
}
\arguments{
\item{dob}{date of birth as a character string "dd/mm/yyyy".}

\item{...}{additional parameters for debugging.}
}
\description{
The final exam for the R-course
}
