% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{section}
\alias{section}
\alias{osio}
\title{Select an exercise section}
\usage{
section(x)

osio(x)
}
\arguments{
\item{x}{an integer value between 1 and 11 to select the section to start.}
}
\description{
Select an exercise section
}
