% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{select_language}
\alias{select_language}
\alias{valitse_kieli}
\title{Set the language used in the section and exam functions.
If no language is specified, the language is switched between English and Finnish.}
\usage{
select_language(language, save_selection)

valitse_kieli(language = c("english", "finnish"), save_selection = FALSE)
}
\arguments{
\item{language}{a character string describing the language to be used (supports "english" or "finnish")}

\item{save_selection}{a logical value that indicates whether the selection should be saved to .Rprofile}
}
\description{
Set the language used in the section and exam functions.
If no language is specified, the language is switched between English and Finnish.
}
