% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/language.R
\name{submit}
\alias{submit}
\alias{vastaa}
\title{Submit an answer to be checked}
\usage{
submit(x)

vastaa(x)
}
\arguments{
\item{x}{an R object to be submitted as the answer}
}
\description{
Submit an answer to be checked
}
